refactor: update NotePlayGround component to save note updates in indexedDB
The NotePlayGround component has been updated to save note updates in indexedDB using the iDBPutNote function. This ensures that changes made to the note content are persisted locally. Additionally, the setUpdated method has been added to the Note class to update the updatedAt property when a note is modified.

Related recent commits:
- feat: use indexed db to save content locally
- feat: Update NoteItem type and refactor related code