add local sync (#9)
* feat: setup tests

* feat: use clamp for fontsizes

* fix(tooling): classname sorting

* feat: setup initial notes ui

* feat(notes): add editor

* refactor(notes): refactor the sidebar to a parallel route

* feat: wip style up a tree component

fix: add padding

* feat([notes- wip]): create note structure

* fix: move nav to layout

- refactore component name

* feat(notes)!: change notes to a map

- chore: mock from backend

* feat: load notes from sidebar

* chore: regenerate lock file

* feat: render tree from server

* fix: grow the layout to the parent's size

* feat: update description in layout.tsx

* chore: update next-themes and zod

* fix: dark theme styles

* chore: debug setup

* feat: get notes from the server and list on file tree

* fix: use client side fetch for notes
feat: add loading state while switching notes

* feat: update API URL in note query to use URL from env

* feat: empty note dashboard page

* chore: remove unneeded package

* feat: note dashboard

* fix: sidebar parallel route

* feat: Update NoteItem type and refactor related code

The NoteItem type has been added to the note module, and the code has been refactored to use this type instead of the Note class.

* feat: use indexed db to save content locally

* refactor: update NotePlayGround component to save note updates in indexedDB

The NotePlayGround component has been updated to save note updates in indexedDB using the iDBPutNote function. This ensures that changes made to the note content are persisted locally. Additionally, the setUpdated method has been added to the Note class to update the updatedAt property when a note is modified.

Related recent commits:
- feat: use indexed db to save content locally
- feat: Update NoteItem type and refactor related code

* refactor: save note contents in every x interval

maybe i could do this for every x keystrokes instead (todo?)

* feat: create new note from the sidebar filetree

* fix: text truncation on the `TreeItem` component

* refactor: add ClickToEdit component to TreeItem for editable node names

* fix: use a defaut note schema for new notes to fix crash when notes are empty

* feat: scroll area for notes on sidebar also use Input to handle name edits

* fix: handle space key press in useEditOnClick hook

The useEditOnClick hook now handles the space key press event to append a space character to the input value. This is a temporary fix until the issue with filetree input handling is resolved.

Related recent commits:
- refactor: add ClickToEdit component to TreeItem for editable node names
- fix: text truncation on the `TreeItem` component

* feat: update note on server

* fix: add missing white space between tree items for better visual hierarchy

* feat: editor sync state indicator

shows whether the document is synced to the indexedDB

* feat: actually implement auto save and manual save and move `useNotes` into the nearest common hook directory

* feat: Add sync button to NotesNav component

This commit adds a sync button to the NotesNav component in order to provide manual saving functionality. The sync button's appearance and behavior are determined by the syncStatus prop, which can have three possible values: "unsynced", "synced", or "saving". The button is disabled when the syncStatus is "synced" or "saving". Clicking the button triggers the manualSave function.

Related recent commits:
- feat: actually implement auto save and manual save and move `useNotes` into the nearest common hook directory
- feat: editor sync state indicator

* chore: Remove unused imports and fix an import path

* refactor: move editor related functionalities (sidebar etc) to a new route group

this prevents rerenders when document is changed from sidebar happens because anything under a dynamic route gets rerendered when the url is changed.

[ref](https://github.com/vercel/next.js/issues/44793#issuecomment-1382458981)